feat: much more refactoring
