+++ .deploy/webapp/Dockerfile
@@ -35,8 +35,6 @@ USER node:node
 
 WORKDIR /srv/gauzy
 
-COPY --chown=node:node apps/desktop/package.json ./apps/desktop/
-COPY --chown=node:node apps/api/package.json ./apps/api/
 COPY --chown=node:node apps/gauzy/package.json ./apps/gauzy/
 
 COPY --chown=node:node packages/common/package.json ./packages/common/
@@ -52,7 +50,7 @@ COPY --chown=node:node packages/plugins/product-reviews/package.json ./packages/
 
 COPY --chown=node:node lerna.json package.json yarn.lock ./
 
-RUN yarn bootstrap
+RUN yarn install --frozen-lockfile && yarn cache clean
 
 FROM node:alpine AS development
 
@@ -89,7 +87,7 @@ ENV DEMO=${DEMO:-false}
 
 ENV IS_DOCKER=true
 
-RUN yarn build:package:all
+RUN yarn build:package:gauzy
 RUN yarn build:gauzy:prod
 
 FROM nginx:alpine AS production
