+++ apps/desktop/package.json
@@ -17,7 +17,9 @@
 		"email": "ever@ever.co",
 		"url": "https://ever.co"
 	},
-	"scripts": {},
+	"scripts": {
+	
+	},
 	"devDependencies": {
 		"electron": "10.1.3",
 		"electron-builder": "^22.9.1",
