+++ apps/gauzy/package.json
@@ -2,5 +2,176 @@
 	"name": "@gauzy/ui",
 	"version": "0.1.0",
 	"description": "Gauzy UI",
-	"license": "AGPL-3.0"
+	"license": "AGPL-3.0",
+	"dependencies": {
+		"@gauzy/common": "^0.1.0",
+		"@gauzy/common-angular": "^0.1.0",
+		"@gauzy/contracts": "^0.1.0",
+		"@agm/core": "^3.0.0-beta.0",
+		"@akveo/ng2-completer": "^9.0.1",
+		"@angular/animations": "^10.1.3",
+		"@angular/cdk": "^10.2.2",
+		"@angular/common": "^10.1.2",
+		"@angular/compiler": "^10.1.3",
+		"@angular/core": "^10.1.3",
+		"@angular/forms": "^10.1.3",
+		"@angular/platform-browser": "^10.1.3",
+		"@angular/platform-browser-dynamic": "^10.1.3",
+		"@angular/router": "^10.1.2",
+		"@angular/service-worker": "^10.1.2",
+		"@asymmetrik/ngx-leaflet": "^8.1.0",
+		"@circlon/angular-tree-component": "^10.0.0",
+		"@cloudinary/angular-5.x": "^1.3.2",
+		"@datorama/akita": "^5.3.0",
+		"@datorama/akita-ngdevtools": "^5.0.3",
+		"@fortawesome/fontawesome-free": "^5.9.0",
+		"@fullcalendar/angular": "^5.3.1",
+		"@fullcalendar/bootstrap": "^5.3.1",
+		"@fullcalendar/core": "^5.3.1",
+		"@fullcalendar/daygrid": "^5.3.2",
+		"@fullcalendar/interaction": "^5.3.1",
+		"@fullcalendar/moment-timezone": "^5.3.1",
+		"@fullcalendar/timegrid": "^5.3.1",
+		"@nebular/auth": "^5.0.0",
+		"@nebular/bootstrap": "^5.0.0",
+		"@nebular/eva-icons": "^5.0.0",
+		"@nebular/security": "^5.0.0",
+		"@nebular/theme": "^5.0.0",
+		"@ng-select/ng-select": "4.0.4",
+		"@ngneat/until-destroy": "^8.0.2",
+		"@ngx-translate/core": "^11.0.1",
+		"@ngx-translate/http-loader": "^4.0.0",
+		"@sentry/angular": "^5.30.0",
+		"@sentry/browser": "^5.30.0",
+		"@sentry/tracing": "^5.30.0",
+		"@swimlane/ngx-charts": "^12.0.1",
+		"angular2-chartjs": "^0.5.1",
+		"angular2-draggable": "^2.3.2",
+		"angular2-toaster": "^8.0.0",
+		"archiver": "^3.1.1",
+		"bcrypt": "^5.0.0",
+		"bootstrap": "^4.3.1",
+		"brace": "^0.11.1",
+		"chart.js": "^2.9.3",
+		"chart.piecelabel.js": "^0.15.0",
+		"ckeditor": "^4.12.1",
+		"ckeditor4-angular": "^2.0.1",
+		"classlist.js": "^1.1.20150312",
+		"cloudinary-core": "^2.7.4",
+		"core-js": "^3.6.5",
+		"csv-parser": "^2.3.2",
+		"csv-writer": "^1.6.0",
+		"d3": "4.13.0",
+		"d3-selection-multi": "1.0.1",
+		"date-fns": "~2.16.1",
+		"date-holidays": "^1.6.1",
+		"detect-passive-events": "^1.0.4",
+		"echarts": "^4.2.1",
+		"email-templates": "^7.1.1",
+		"eva-icons": "^1.1.3",
+		"file-saver": "^2.0.2",
+		"filepond": "^4.5.1",
+		"filepond-plugin-file-encode": "^2.1.0",
+		"filepond-plugin-file-validate-size": "^2.2.0",
+		"filepond-plugin-file-validate-type": "^1.2.0",
+		"filepond-plugin-image-crop": "^2.0.0",
+		"filepond-plugin-image-exif-orientation": "^1.0.0",
+		"filepond-plugin-image-preview": "^4.4.0",
+		"filepond-plugin-image-resize": "^2.0.0",
+		"filepond-plugin-image-transform": "^3.4.3",
+		"formidable": "^1.2.2",
+		"fullcalendar": "^5.3.0",
+		"hammerjs": "^2.0.8",
+		"handlebars": "^4.7.6",
+		"html-to-text": "^6.0.0",
+		"html2canvas": "^1.0.0-rc.7",
+		"immer": "^8.0.1",
+		"intl": "^1.2.5",
+		"ionicons": "^4.6.3",
+		"jsonwebtoken": "^8.5.1",
+		"keycloak-js": "^11.0.2",
+		"leaflet": "^1.7.1",
+		"moment": "^2.29.0",
+		"moment-range": "^4.0.2",
+		"moment-timezone": "^0.5.31",
+		"nebular-icons": "^1.1.0",
+		"ng2-ace-editor": "^0.3.7",
+		"ng2-ckeditor": "^1.2.6",
+		"ng2-file-upload": "^1.3.0",
+		"ng2-smart-table": "^1.6.0",
+		"ng5-slider": "^1.2.4",
+		"ngx-color-picker": "^8.2.0",
+		"ngx-echarts": "^4.2.1",
+		"ngx-electron": "^2.2.0",
+		"ngx-feature-toggle": "^10.0.0",
+		"ngx-filepond": "^5.0.1",
+		"ngx-moment": "^5.0.0",
+		"ngx-page-scroll": "^6.0.2",
+		"ngx-page-scroll-core": "^6.0.2",
+		"ngx-perfect-scrollbar": "^8.0.0",
+		"ngx-permissions": "^8.1.1",
+		"normalize.css": "^8.0.1",
+		"pace-js": "^1.0.2",
+		"pdfmake": "^0.1.70",
+		"pug": "^2.0.4",
+		"reflect-metadata": "^0.1.13",
+		"roboto-fontface": "^0.10.0",
+		"rxjs": "^6.6.3",
+		"rxjs-compat": "^6.6.3",
+		"screenfull": "^5.0.2",
+		"smooth-scrollbar": "^8.5.3",
+		"socicon": "^3.0.5",
+		"sqlite3": "https://github.com/ever-co/node-sqlite3",
+		"squirrelly": "^8.0.8",
+		"tinymce": "^5.4.2",
+		"tslib": "^2.0.1",
+		"twing": "^5.0.2",
+		"typeface-exo": "^0.0.61",
+		"underscore": "^1.12.0",
+		"unleash": "^2.0.2",
+		"unzipper": "^0.10.11",
+		"uuid": "^8.3.0",
+		"web-animations-js": "^2.3.2",
+		"yargs": "^16.2.0",
+		"zone.js": "~0.11.1"
+	},
+	"devDependencies": {
+		"@angular-builders/custom-webpack": "^10.0.1",
+		"@angular-devkit/build-angular": "~0.1001.3",
+		"@angular-devkit/build-ng-packagr": "~0.1001.3",
+		"@angular/cli": "^10.1.3",
+		"@angular/compiler-cli": "^10.1.3",
+		"@angular/language-service": "^10.1.3",
+		"@types/async": "^3.2.3",
+		"@types/bcrypt": "^3.0.0",
+		"@types/chart.js": "^2.9.15",
+		"@types/d3-color": "^1.2.2",
+		"@types/email-templates": "^7.1.0",
+		"@types/faker": "^4.1.12",
+		"@types/google.analytics": "^0.0.40",
+		"@types/googlemaps": "^3.36.5",
+		"@types/hammerjs": "^2.0.36",
+		"@types/html-to-text": "5.1.0",
+		"@types/i18n": "^0.8.6",
+		"@types/jest": "^24.0.15",
+		"@types/leaflet": "^1.5.19",
+		"@types/multer": "^1.4.4",
+		"@types/multer-s3": "^2.7.7",
+		"@types/node": "^14.14.22",
+		"@types/underscore": "^1.9.4",
+		"@types/uuid": "^3.4.4",
+		"@types/yargs": "^15.0.9",
+		"cypress": "^5.2.0",
+		"jasmine-core": "^3.6.0",
+		"jasmine-spec-reporter": "^6.0.0",
+		"jest": "^26.4.2",
+		"jest-preset-angular": "^8.3.0",
+		"karma": "^5.2.3",
+		"karma-chrome-launcher": "^3.1.0",
+		"karma-cli": "^2.0.0",
+		"karma-coverage-istanbul-reporter": "^3.0.3",
+		"karma-jasmine": "^4.0.1",
+		"karma-jasmine-html-reporter": "^1.5.4",
+		"pm2": "^4.5.1"
+	}
 }
