+++ apps/gauzy/src/app/@core/services/employees.service.ts
@@ -8,7 +8,7 @@ import {
 } from '@gauzy/contracts';
 import { Observable } from 'rxjs';
 import { first } from 'rxjs/operators';
-import { toParams } from '@gauzy/common';
+import { toParams } from '@gauzy/common-angular';
 
 @Injectable()
 export class EmployeesService {
