+++ apps/gauzy/src/app/@core/services/job-preset.service.ts
@@ -6,7 +6,7 @@ import {
 	IJobPreset,
 	IMatchingCriterions
 } from '@gauzy/contracts';
-import { toParams } from '@gauzy/common';
+import { toParams } from '@gauzy/common-angular';
 
 @Injectable({
 	providedIn: 'root'
