+++ apps/gauzy/src/app/@core/services/organization-contact.service.ts
@@ -7,7 +7,7 @@ import {
 	IEditEntityByMemberInput
 } from '@gauzy/contracts';
 import { first } from 'rxjs/operators';
-import { toParams } from '@gauzy/common';
+import { toParams } from '@gauzy/common-angular';
 
 @Injectable({
 	providedIn: 'root'
