+++ apps/gauzy/src/app/@core/services/payment.service.ts
@@ -9,7 +9,7 @@ import {
 	IPaymentUpdateInput
 } from '@gauzy/contracts';
 import { first } from 'rxjs/operators';
-import { toParams } from '@gauzy/common';
+import { toParams } from '@gauzy/common-angular';
 
 @Injectable({
 	providedIn: 'root'
