+++ apps/gauzy/src/app/@shared/timesheet/activity.service.ts
@@ -5,7 +5,7 @@ import {
 	IGetActivitiesInput,
 	IDailyActivity
 } from '@gauzy/contracts';
-import { toParams } from '@gauzy/common';
+import { toParams } from '@gauzy/common-angular';
 
 @Injectable({
 	providedIn: 'root'
