+++ apps/gauzy/src/app/@shared/timesheet/timesheet-statistics.service.ts
@@ -15,7 +15,7 @@ import {
 	ITasksStatistics,
 	IManualTimesStatistics
 } from '@gauzy/contracts';
-import { toParams } from '@gauzy/common';
+import { toParams } from '@gauzy/common-angular';
 
 @Injectable({
 	providedIn: 'root'
