+++ packages/common/src/utils/shared-utils.ts
@@ -1,6 +1,5 @@
 import * as moment from 'moment';
 import * as timezone from 'moment-timezone';
-import { HttpParams } from '@angular/common/http';
 
 /**
  * Check is function .
@@ -149,35 +148,6 @@ export function progressStatus(value) {
 	} else {
 		return 'danger';
 	}
-}
-
-// It will use for pass nested object or array in query params in get method.
-export function toParams(query) {
-	let params: HttpParams = new HttpParams();
-	Object.keys(query).forEach((key) => {
-		if (isJsObject(query[key])) {
-			params = toSubParams(params, key, query[key]);
-		} else {
-			params = params.append(key.toString(), query[key]);
-		}
-	});
-	return params;
-}
-
-function toSubParams(params: HttpParams, key: string, object: any) {
-	Object.keys(object).forEach((childKey) => {
-		if (isJsObject(object[childKey])) {
-			params = toSubParams(
-				params,
-				`${key}[${childKey}]`,
-				object[childKey]
-			);
-		} else {
-			params = params.append(`${key}[${childKey}]`, object[childKey]);
-		}
-	});
-
-	return params;
 }
 
 export function isJsObject(object: any) {
