+++ packages/contracts/src/country.model.ts
@@ -1,4 +1,5 @@
 import { IBaseEntityModel } from './base-entity.model';
+
 export interface ICountry extends IBaseEntityModel {
 	isoCode: string;
 	country: string;
