+++ packages/contracts/src/invite.model.ts
@@ -1,7 +1,3 @@
-// Modified code from https://github.com/xmlking/ngx-starter-kit.
-// MIT License, see https://github.com/xmlking/ngx-starter-kit/blob/develop/LICENSE
-// Copyright (c) 2018 Sumanth Chinthagunta
-
 import { IRole } from './role.model';
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
 import { IUser, IUserRegistrationInput, LanguagesEnum } from './user.model';
