+++ packages/core/package.json
@@ -86,6 +86,7 @@
 		"node-fetch": "^2.6.1",
 		"nodemailer": "^6.4.11",
 		"nodemailer-handlebars": "^1.0.1",
+		"date-fns": "~2.16.1",
 		"passport": "^0.4.1",
 		"passport-auth0": "^1.3.3",
 		"passport-azure-ad-oauth2": "^0.0.4",
